Reorganize RC branch in anticipate of Oct release (#287)
We now have two different prefixes that append to the data plane API :
`bulk` and `vectors`. As such, it's time to reorganize our repo!

The `data/` dir now is organized into 2 subdirs: `bulk/` and `vectors/`.
Tests share the same organization.

`data/index.ts` is still at the `data/` dir level (as is
`indexHostSingleton.ts`), since that's the entry point for both prefixes
to reach the data plane (i.e. the `Index` class); tests reflect this as
well.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

CI passes.
