+++ src/control/indexOperationsBuilder.ts
@@ -10,7 +10,7 @@ import {
   normalizeUrl,
 } from '../utils';
 import { middleware } from '../utils/middleware';
-import type { PineconeConfiguration } from '../data/types';
+import type { PineconeConfiguration } from '../data/vectors/types';
 import type { ConfigurationParameters as IndexOperationsApiConfigurationParameters } from '../pinecone-generated-ts-fetch/db_control';
 
 export const indexOperationsBuilder = (
