+++ src/data/__tests__/vectors/deleteOne.test.ts
@@ -1,10 +1,10 @@
-import { deleteOne } from '../deleteOne';
+import { deleteOne } from '../../vectors/deleteOne';
 import type {
   DeleteOperationRequest,
   VectorOperationsApi as DataPlaneApi,
-} from '../../pinecone-generated-ts-fetch/db_data';
-import { DataOperationsProvider } from '../dataOperationsProvider';
-import { PineconeArgumentError } from '../../errors';
+} from '../../../pinecone-generated-ts-fetch/db_data';
+import { DataOperationsProvider } from '../../vectors/dataOperationsProvider';
+import { PineconeArgumentError } from '../../../errors';
 
 const setupDeleteResponse = (response, isSuccess) => {
   const fakeDelete: (req: DeleteOperationRequest) => Promise<object> = jest
