+++ src/data/__tests__/vectors/describeIndexStats.test.ts
@@ -1,8 +1,8 @@
-import { describeIndexStats } from '../describeIndexStats';
-import { VectorOperationsApi as DataPlaneApi } from '../../pinecone-generated-ts-fetch/db_data';
-import { DataOperationsProvider } from '../dataOperationsProvider';
-import type { DescribeIndexStatsOperationRequest } from '../../pinecone-generated-ts-fetch/db_data';
-import { PineconeArgumentError } from '../../errors';
+import { describeIndexStats } from '../../vectors/describeIndexStats';
+import { VectorOperationsApi as DataPlaneApi } from '../../../pinecone-generated-ts-fetch/db_data';
+import { DataOperationsProvider } from '../../vectors/dataOperationsProvider';
+import type { DescribeIndexStatsOperationRequest } from '../../../pinecone-generated-ts-fetch/db_data';
+import { PineconeArgumentError } from '../../../errors';
 
 const setupResponse = (response, isSuccess) => {
   const fakeDescribeIndexStats: (
