+++ src/data/__tests__/vectors/update.test.ts
@@ -1,7 +1,7 @@
-import { UpdateCommand } from '../update';
-import { VectorOperationsApi as DataPlaneApi } from '../../pinecone-generated-ts-fetch/db_data';
-import { DataOperationsProvider } from '../dataOperationsProvider';
-import type { UpdateOperationRequest } from '../../pinecone-generated-ts-fetch/db_data';
+import { UpdateCommand } from '../../vectors/update';
+import { VectorOperationsApi as DataPlaneApi } from '../../../pinecone-generated-ts-fetch/db_data';
+import { DataOperationsProvider } from '../../vectors/dataOperationsProvider';
+import type { UpdateOperationRequest } from '../../../pinecone-generated-ts-fetch/db_data';
 
 const setupResponse = (response, isSuccess) => {
   const fakeUpdate: (req: UpdateOperationRequest) => Promise<object> = jest
