+++ src/data/vectors/deleteAll.ts
