+++ src/data/vectors/deleteMany.ts
@@ -1,7 +1,7 @@
 import { DataOperationsProvider } from './dataOperationsProvider';
-import type { DeleteRequest } from '../pinecone-generated-ts-fetch/db_data';
+import type { DeleteRequest } from '../../pinecone-generated-ts-fetch/db_data';
 import type { RecordId } from './types';
-import { PineconeArgumentError } from '../errors';
+import { PineconeArgumentError } from '../../errors';
 
 /**
  * A list of record ids to delete from the index.
