+++ src/integration/data/vectors/fetch.test.ts
@@ -1,5 +1,10 @@
-import { Pinecone, Index } from '../../index';
-import { globalNamespaceOne, getRecordIds } from '../test-helpers';
+import { Pinecone, Index } from '../../../index';
+import {
+  generateRecords,
+  randomString,
+  INDEX_NAME,
+  waitUntilRecordsReady,
+} from '../../test-helpers';
 
 // todo: add pod tests
