+++ src/integration/data/vectors/query.test.ts
@@ -1,9 +1,9 @@
-import { Index, Pinecone, QueryResponse } from '../../index';
+import { Pinecone, Index } from '../../../index';
 import {
   globalNamespaceOne,
   getRecordIds,
   assertWithRetries,
-} from '../test-helpers';
+} from '../../test-helpers';
 
 let pinecone: Pinecone,
   serverlessIndex: Index,
