+++ src/pinecone.ts
@@ -29,7 +29,7 @@ import { Inference } from './inference';
 import { inferenceOperationsBuilder } from './inference/inferenceOperationsBuilder';
 import { isBrowser } from './utils/environment';
 import { ValidateProperties } from './utils/validateProperties';
-import { PineconeConfigurationProperties } from './data/types';
+import { PineconeConfigurationProperties } from './data/vectors/types';
 
 /**
  * The `Pinecone` class is the main entrypoint to this sdk. You will use
